+++ tests/testthat/test_celltype_specificity_Aif1_Pvalb.R
@@ -1,6 +1,6 @@
 # Test Cell type specificity calculations Aif1, Pvalb
 test_that("Cell type specificity Aif1, Pvalb", {
-    # inspect specificity matrix for type 1 cell types from generate.celltype.data
+    # inspect specificity matrix for type 1 cell types from generate_celltype_data
     # Use Vignette Dataset to check function output
     cortex_mrna <- cortex_mrna()
    # eh <- query(ExperimentHub::ExperimentHub(), "ewceData")
@@ -17,18 +17,18 @@ test_that("Cell type specificity Aif1, Pvalb", {
 
     # Note not normalising data to save time
 
-    exp_CortexOnly_DROPPED <- drop.uninformative.genes(
+    exp_CortexOnly_DROPPED <- drop_uninformative_genes(
         exp = cortex_mrna$exp, # _scT_normed,
         level2annot = cortex_mrna$annot$level2class
     )
     annotLevels <- list(level1class = cortex_mrna$annot$level1class, level2class = cortex_mrna$annot$level2class)
-    fNames_CortexOnly <- generate.celltype.data(exp = exp_CortexOnly_DROPPED, 
+    fNames_CortexOnly <- generate_celltype_data(exp = exp_CortexOnly_DROPPED, 
                                                     annotLevels = annotLevels, 
                                                     groupName = "kiCortexOnly",
                                                     savePath = tempdir())
 
     # filter only orthologs
-    fNames_CortexOnly <- filter.genes.without.1to1.homolog(fNames_CortexOnly)
+    fNames_CortexOnly <- filter_genes_without_1to1_homolog(fNames_CortexOnly)
 
     # load and inspect specificity matrix
     load(fNames_CortexOnly[1])
@@ -42,12 +42,12 @@ test_that("Cell type specificity Aif1, Pvalb", {
     # Try running with rank norm transformation normSpec set to True
     # Ensure same most specific cell types remain
     ctd_not_norm_spec <- ctd
-    fNames_CortexOnly <- generate.celltype.data(
+    fNames_CortexOnly <- generate_celltype_data(
         exp = exp_CortexOnly_DROPPED, annotLevels = annotLevels,
         groupName = "kiCortexOnly", normSpec = TRUE, savePath=tempdir()
     )
     # filter only orthologs
-    fNames_CortexOnly <- filter.genes.without.1to1.homolog(fNames_CortexOnly)
+    fNames_CortexOnly <- filter_genes_without_1to1_homolog(fNames_CortexOnly)
 
     # load and inspect specificity matrix
     load(fNames_CortexOnly[1])