+++ tests/testthat/test_fix_bad_mgi_symbols.r
@@ -1,4 +1,4 @@
-# Test for fix.bad.mgi.symbols
+# Test for fix_bad_mgi_symbols
 test_that("method to remove/fix an expected set of genes", {
     # Use Vignette Dataset to check function, alter input gene names
     cortex_mrna <- cortex_mrna()
@@ -18,7 +18,7 @@ test_that("method to remove/fix an expected set of genes", {
 
     # catch error when no exp inputted
     error_return <-
-        tryCatch(EWCE::fix.bad.mgi.symbols(
+        tryCatch(EWCE::fix_bad_mgi_symbols(
             exp = NULL,
             mrk_file_path = sprintf("%s/MRK_List2.rpt", tempdir())
         ),
@@ -30,7 +30,7 @@ test_that("method to remove/fix an expected set of genes", {
     test_exp_set_char <- apply(test_exp_set, 2, as.character)
     rownames(test_exp_set_char) <- rownames(test_exp_set)
     error_return2 <-
-        tryCatch(EWCE::fix.bad.mgi.symbols(
+        tryCatch(EWCE::fix_bad_mgi_symbols(
             exp = test_exp_set_char,
             mrk_file_path = sprintf("%s/MRK_List2.rpt", tempdir())
         ),
@@ -40,7 +40,7 @@ test_that("method to remove/fix an expected set of genes", {
 
     # pass data table rather than data frame or matrix
     error_return3 <-
-        tryCatch(EWCE::fix.bad.mgi.symbols(
+        tryCatch(EWCE::fix_bad_mgi_symbols(
             exp = data.table::data.table(as.data.frame(test_exp_set)),
             mrk_file_path = sprintf("%s/MRK_List2.rpt", tempdir())
         ),
@@ -50,7 +50,7 @@ test_that("method to remove/fix an expected set of genes", {
 
     # function should warn the user about this -if warning returned function worked
     warning_return <-
-        tryCatch(EWCE::fix.bad.mgi.symbols(test_exp_set,
+        tryCatch(EWCE::fix_bad_mgi_symbols(test_exp_set,
             mrk_file_path = sprintf("%s/MRK_List2.rpt", tempdir())
         ),
         error = function(e) e,
@@ -59,27 +59,27 @@ test_that("method to remove/fix an expected set of genes", {
 
     # running on hgnc rather than mgi should return warnings
     warning_return2 <-
-        tryCatch(EWCE::fix.bad.hgnc.symbols(test_exp_set),
+        tryCatch(EWCE::fix_bad_hgnc_symbols(test_exp_set),
             error = function(e) e,
             warning = function(w) w
         )
 
     # running on hgnc rather than mgi should return warnings
     warning_return3 <-
-        tryCatch(EWCE::fix.bad.hgnc.symbols(data.table::data.table(as.data.frame(test_exp_set))),
+        tryCatch(EWCE::fix_bad_hgnc_symbols(data.table::data.table(as.data.frame(test_exp_set))),
             error = function(e) e,
             warning = function(w) w
         )
 
     options(warn = -1)
-    hgnc_return <- EWCE::fix.bad.hgnc.symbols(test_exp_set)
+    hgnc_return <- EWCE::fix_bad_hgnc_symbols(test_exp_set)
     options(warn = 0)
 
     # Now test if a synonym of a gene in the list is added
     # function should combine them and give sum reads for each sample
     # alt symbol for Tspan12 is Tm4sf12
     rownames(test_exp_set)[7] <- "Tm4sf12"
-    EWCE_return <- EWCE::fix.bad.mgi.symbols(test_exp_set,
+    EWCE_return <- EWCE::fix_bad_mgi_symbols(test_exp_set,
         mrk_file_path = sprintf("%s/MRK_List2.rpt", tempdir()),
         printAllBadSymbols = TRUE
     )
@@ -89,7 +89,7 @@ test_that("method to remove/fix an expected set of genes", {
 
     # check nothing changes when there are no issues
     test_exp_set <- test_exp_set[1:5, ]
-    EWCE_output_same_input <- EWCE::fix.bad.mgi.symbols(test_exp_set,
+    EWCE_output_same_input <- EWCE::fix_bad_mgi_symbols(test_exp_set,
         mrk_file_path = sprintf("%s/MRK_List2.rpt", tempdir())
     )
 
@@ -97,7 +97,7 @@ test_that("method to remove/fix an expected set of genes", {
     # check input runs on large size of incorrect mgi gene names
     # check it catches warning still but doesn't give an error
     warning_return4 <-
-        tryCatch(EWCE_output_large <- EWCE::fix.bad.mgi.symbols(cortex_mrna$exp[1:10000, 1:1000],
+        tryCatch(EWCE_output_large <- EWCE::fix_bad_mgi_symbols(cortex_mrna$exp[1:10000, 1:1000],
             mrk_file_path = sprintf("%s/MRK_List2.rpt", tempdir())
         ),
         error = function(e) e,