+++ NEWS.md
@@ -5,7 +5,7 @@
 
     *   EWCE v1.0 on Bioconductor replaces the defunct [EWCE v1.3.0](https://bioconductor.riken.jp/packages/3.5/bioc/html/EWCE.html) available on Bioconductor v3.5.
     *   EWCE has been rendered scalable to the analysis of large datasets
-    *   `drop.uninformative.genes()` has been expanded to allow the utilisation of differential expression approaches 
+    *   `drop_uninformative_genes()` has been expanded to allow the utilisation of differential expression approaches 
 
 
 ## USER-VISIBLE CHANGES
@@ -15,7 +15,27 @@
 
 ## DEPRECATED & DEFUNCT
 
-    *   None
+    *   The following functions have been renamed to use underscore in compliance with Bioconductor nomenclature:
+        check.ewce.genelist.inputs
+        cell.list.dist
+        bootstrap.enrichment.test
+        bin.specificity.into.quantiles
+        bin.columns.into.quantiles
+        add.res.to.merging.list
+        prepare.genesize.control.network
+        prep.dendro 
+        get.celltype.table
+        calculate.specificity.for.level
+        calculate.meanexp.for.level 
+        generate.celltype.data
+        generate.bootstrap.plots
+        generate.bootstrap.plots.for.transcriptome
+        fix.bad.mgi.symbols
+        fix.bad.hgnc.symbols
+        filter.genes.without.1to1.homolog
+        ewce.plot
+        cells.in.ctd
+        drop.uninformative.genes
 
 
 ## BUG FIXES